feat: Create component table to store labels
There are two important parts to defining components
in an ecs system: the component type and the actual
components. The component type, which I will refer
to as the component, stores the labels to the actual
components as well as some basic information to describe
the component in english and with a default state.